additional file separation + renaming files
